Refactor setup scripts
